Move function from plot_utilities to core_utilities
